% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStatistic.R
\name{changepoint_variance}
\alias{changepoint_variance}
\title{Compute the variance across a changepoint model}
\usage{
changepoint_variance(y, fit, ...)
}
\arguments{
\item{y}{data vector}

\item{fit}{fitted changepoint object}

\item{...}{not used}
}
\value{
numeric
}
\description{
Compute the variance across a changepoint model
}
