% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declutter.R
\name{contrast_from_cluster}
\alias{contrast_from_cluster}
\title{Construct the contrast vector from \code{declutter}}
\usage{
contrast_from_cluster(cluster_list, n, location)
}
\arguments{
\item{cluster_list}{The output from \code{declutter} function}

\item{n}{Positive integer, for length of contrast}

\item{location}{positive index, corresponding to the jump location in \code{cluster_list$jump_vec}}
}
\value{
vector of length \code{n}
}
\description{
Construct the contrast vector from \code{declutter}
}
