% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declutter.R
\name{declutter}
\alias{declutter}
\title{Declutter of jumps}
\usage{
declutter(jump_vec, sign_vec, how_close = 3, desired_jumps = NA)
}
\arguments{
\item{jump_vec}{vector of positive integers}

\item{sign_vec}{vector containing \code{c(-1,1)} (possibly repeating) of
same length as \code{jump_vec}}

\item{how_close}{positive integer}

\item{desired_jumps}{vector of positive integers}
}
\value{
A list containing \code{jump_vec} (the vector of median jumps),
\code{sign_mat}, a matrix counting the signs within each cluster, and
\code{target_bool}, a boolean vector stating which clusters "contained"
\code{desired_jumps}
}
\description{
Takes in a jump vector \code{jump_vec} as well as a vector of signs
\code{sign_vec} (of equal length). Then, based on \code{how_close},
\code{declutter} finds clusters of jumps as well as returns the
signs of each cluster's members.
}
\details{
Also, \code{desired_jumps} is an input that informs the output on
which jump cluster "contained" the desired jump locations. This is passed in
as a vector. This is stated in the output \code{target_bool}
}
