% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{.conditional_gaussian}
\alias{.conditional_gaussian}
\title{Compute the conditional density of a Gaussian distribution}
\source{
\url{https://en.wikipedia.org/wiki/Multivariate_normal_distribution#Conditional_distributions}
}
\usage{
.conditional_gaussian(gaussian, val)
}
\arguments{
\item{gaussian}{a \code{gaussian} object}

\item{val}{the values of the last coordinates}
}
\value{
a \code{gaussian} object
}
\description{
If \code{val} is length \code{k} long, then this function returns the
\code{gaussian} object corresponding to density conditioned on the last
\code{k} values equal to \code{val}. For example, if \code{gaussian} represents
a 5-dimensional distribution, and \code{val = c(5,10)}, then this function
returns a \code{gaussian} object corresponding to the Gaussian distribution
where the 4th coordinate is set to 5 and the 5th coordinate is set to 10.
}
