% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitrun_radial.R
\name{.hit_run_next_point_radial}
\alias{.hit_run_next_point_radial}
\title{Output the next point for hit-and-run sampler, Radial}
\usage{
.hit_run_next_point_radial(y, null_mat, polyhedra)
}
\arguments{
\item{y}{data}

\item{null_mat}{matrix to represent the null space of \code{.segments}}

\item{polyhedra}{\code{polyhedra} object}
}
\value{
vector
}
\description{
For unknown sigma
}
