% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStatistic.R
\name{.jump_contrast}
\alias{.jump_contrast}
\title{Compute distance of next jump}
\usage{
.jump_contrast(y, jump_vec_before, jump_vec_after)
}
\arguments{
\item{y}{data}

\item{jump_vec_before}{vector of indices}

\item{jump_vec_after}{vector of indices with exactly one additional element
compared to \code{jump_vec_before}}
}
\value{
positive numeric
}
\description{
\code{jump_vec} is a vector of indices (not necessarily in any particular
order) less than or equal to \code{n}. A value of \code{i} in \code{jump_vec}
denotes that there is jump between index \code{i} and \code{i+1}.
\code{jump_vec} does not need to include \code{n}.
}
