% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{.projection_matrix}
\alias{.projection_matrix}
\title{Projection of vector onto rows of a matrix}
\usage{
.projection_matrix(vec, mat)
}
\arguments{
\item{vec}{vector}

\item{mat}{matrix}
}
\value{
vector
}
\description{
Returns the component of \code{vec} that is orthogonal to all the rows
of \code{mat}
}
