% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radians.R
\name{.radians_to_data}
\alias{.radians_to_data}
\title{Convert radians into a sample}
\usage{
.radians_to_data(theta, y, v, w)
}
\arguments{
\item{theta}{radians}

\item{y}{vector}

\item{v}{unit vector}

\item{w}{unit vector orthogonal to \code{v}}
}
\value{
vector with length of \code{length(y)}
}
\description{
Convert radians into a sample
}
