% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radians.R
\name{.radius}
\alias{.radius}
\title{Radius function}
\usage{
.radius(theta, y, v, w)
}
\arguments{
\item{theta}{radians}

\item{y}{vector}

\item{v}{unit vector}

\item{w}{unit vector orthogonal to \code{v}}
}
\value{
numeric
}
\description{
Radius function
}
