% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radians.R
\name{.range_theta_polyhedra}
\alias{.range_theta_polyhedra}
\title{Find the range of theta's that are in the polyhedra}
\usage{
.range_theta_polyhedra(y, v, w, polyhedra)
}
\arguments{
\item{y}{vector}

\item{v}{unit vector}

\item{w}{unit vector orthogonal to \code{v}}

\item{polyhedra}{\code{polyhedra} object}
}
\value{
vector of possible theta's
}
\description{
Finds theta's such that for each theta,
\code{.radians_to_data(theta, y, v, w)} is inside the polyhedra.
}
