% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{.sample_matrix_space}
\alias{.sample_matrix_space}
\title{Sample unit vectors from the null space of a matrix}
\usage{
.sample_matrix_space(mat, num_vec = NA, null = T)
}
\arguments{
\item{mat}{matrix}

\item{num_vec}{positive integer or \code{NA}}

\item{null}{boolean}
}
\value{
a matrix of vectors, with number of rows equal to \code{ncol(mat)}
and number of columns equal to \code{num_vec}
}
\description{
This function assumes the row space of \code{mat} is the number of
rows in \code{mat}. If \code{num_vec} is \code{NA}, then the function
construct a basis for the entire null space of \code{mat}, which is
assumed to have dimension equal to the rank of \code{mat}.
}
\details{
If \code{null} is \code{TRUE}, then this function samples from the nullspace.
Otherwise, sample from the rowspace.
}
