% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitrun_radial.R
\name{.sampler_hit_run_radial}
\alias{.sampler_hit_run_radial}
\title{Hit and run sampler, Radial}
\usage{
.sampler_hit_run_radial(y, segments, polyhedra, num_samp = 100,
  burn_in = 500, lapse = 2, verbose = F)
}
\arguments{
\item{y}{data}

\item{segments}{matrix created by \code{.segments}}

\item{polyhedra}{\code{polyhedra} object}

\item{num_samp}{number of desired samples from null distribution}

\item{burn_in}{positive integer of the first few samples to throw out per core}

\item{lapse}{positive integer, where we sample \code{num_samp*burn_in}
samples from the null distribution and return every \code{burn_in}th sample}

\item{verbose}{boolean}
}
\value{
matrix with \code{num_samp} columns and \code{length(y)} rows
}
\description{
For unknown sigma
}
