% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectiveModel.R
\name{.segment_means}
\alias{.segment_means}
\title{Computes the empirical mean of each segment}
\usage{
.segment_means(y, segments)
}
\arguments{
\item{y}{data}

\item{segments}{matrix created by \code{.segments}}
}
\value{
vector
}
\description{
Computes the empirical mean of each segment
}
