% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectiveModel.R
\name{.segments}
\alias{.segments}
\title{Compute the matrix of segments}
\usage{
.segments(n, jump_vec, ignore_jump = NA)
}
\arguments{
\item{n}{number of samples}

\item{jump_vec}{vector of indices}

\item{ignore_jump}{which jump to ignore}
}
\value{
s-row matrix, where s is equal to \code{length(jump_vec+1)}
}
\description{
Compute the matrix of segments
}
\details{
\code{jump_vec} is a vector of indices (not necessarily in any particular
order) less than or equal to \code{n}. A value of \code{i} in \code{jump_vec}
denotes that there is jump between index \code{i} and \code{i+1}.
\code{jump_vec} does not need to include \code{n}.
}
