% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radians.R
\name{.try_polyhedra}
\alias{.try_polyhedra}
\title{Determine if each y is in polyhedra}
\usage{
.try_polyhedra(y_mat, polyhedra)
}
\arguments{
\item{y_mat}{matrix where each column represents a different sample, or
a vector}

\item{polyhedra}{\code{polyhedra} object}
}
\value{
vector of booleans, one for each column of \code{y_mat} (or just one)
}
\description{
Determine if each y is in polyhedra
}
