% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStatistic.R
\name{next_jump.bsFs}
\alias{next_jump.bsFs}
\title{Compute the next jump for a \code{bsFs} object}
\usage{
next_jump.bsFs(obj, y, ...)
}
\arguments{
\item{obj}{\code{bsFs} object}

\item{y}{data}

\item{...}{additional parameters}
}
\value{
\code{bsFs} object with one more jump
}
\description{
Compute the next jump for a \code{bsFs} object
}
