% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{next_jump}
\alias{next_jump}
\title{Compute next jump}
\usage{
next_jump(obj, y, ...)
}
\arguments{
\item{obj}{object}

\item{y}{data}

\item{...}{additional parameters}
}
\value{
object with one more jump
}
\description{
Compute next jump
}
