% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStatistic.R
\name{next_jump_statistic}
\alias{next_jump_statistic}
\title{Test statistic for testing empirical size of next jump}
\usage{
next_jump_statistic(y, fit, ...)
}
\arguments{
\item{y}{data}

\item{fit}{changepoint fitted model}

\item{...}{nothing}
}
\value{
numeric
}
\description{
Test statistic for testing empirical size of next jump
}
