% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStatistic.R
\name{segment_difference}
\alias{segment_difference}
\title{Compute the height of a specific jump}
\usage{
segment_difference(y, fit, jump = 1, ...)
}
\arguments{
\item{y}{data vector}

\item{fit}{fitted changepoint object}

\item{jump}{positive integer}

\item{...}{not used}
}
\value{
numeric
}
\description{
\code{jump_number} refers to the \code{jump_number}th jump starting from
the left. This function outputs the difference between the mean of \code{y}
to the left and to the right of that jump (right minus left)
}
