% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectiveModel.R
\name{selected_model_inference}
\alias{selected_model_inference}
\title{Selected model inference}
\usage{
selected_model_inference(y, fit_method, test_func = segment_difference,
  declutter_func = function(x) {     x }, null_mean = rep(0,
  length(y)), num_samp = 100, sigma = NA, ignore_jump = 1,
  direction = NA, param = list(burn_in = default_burn_in(), lapse =
  default_lapse()), return_samples = F, verbose = T, ...)
}
\arguments{
\item{y}{data of length n}

\item{fit_method}{function to fit changepoint model}

\item{test_func}{function to apply to each \code{y} and fitted model}

\item{declutter_func}{function to apply to the vector of fitted jumps}

\item{null_mean}{null mean vector}

\item{num_samp}{number of desired samples from null distribution}

\item{sigma}{postive number or \code{NA}.}

\item{ignore_jump}{which jump to ignore}

\item{direction}{either NA (two-sided test) or -1 or 1 (for a one-sided test)}

\item{param}{additional parameters for \code{sample_method} passed in as a list}

\item{return_samples}{boolean on whether or not the null samples are returned. If
\code{TRUE}, an additional element of the result will be returned}

\item{verbose}{boolean}

\item{...}{optional inputs for \code{test_func}}
}
\value{
a list
}
\description{
Fits a changepoint model using \code{fit_method} to \code{y}. The returned
model needs to have a function defined via the generic function
\code{binSeg::polyhedra}. Inference is performed by sampling data from the
null distribution matching the estimated changepoint model
where each sample is drawn from a Gaussian and matches
in empirical means in each segment and in L2-norm (the sufficient statistics).
}
\details{
Currently the test is a 2-sided test.

The \code{test_func} should be a function that takes in \code{y} as the first
input and the output of \code{fit_method} as the second input, and \code{...}
as the third input.

If \code{sigma} is \code{NA}, then the function conditions on \code{.l2norm(y)}.
Otherwise, \code{sigma} represents the standard deviation of the marginal distribution
of \code{y}, asserting that each coordinate is independent and Gaussian.
}
