% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectiveModel.R
\name{selected_model_inference}
\alias{selected_model_inference}
\title{Selected model inference}
\usage{
selected_model_inference(y, fit_method, test_func = next_jump_statistic,
  num_samp = 100, sample_method = "hitrun", sigma = NA,
  ignore_jump = NA, param = list(burn_in = default_burn_in(), lapse =
  default_lapse(), time_limit = default_time_limit()), cores = 1,
  verbose = T, ...)
}
\arguments{
\item{y}{data of length n}

\item{fit_method}{function to fit changepoint model}

\item{test_func}{function to apply to each \code{y} and fitted model}

\item{num_samp}{number of desired samples from null distribution}

\item{sample_method}{either \code{hit_run} or \code{rejection}}

\item{sigma}{postive number or \code{NA}.}

\item{ignore_jump}{which jump to ignore}

\item{param}{additional parameters for \code{sample_method} passed in as a list}

\item{cores}{number of cores}

\item{verbose}{boolean}

\item{...}{optional inputs for \code{test_func}}
}
\value{
quantile
}
\description{
Fits a changepoint model using \code{fit_method} to \code{y}. The returned
model needs to have a function defined via the generic function
\code{binSeg::polyhedra}. Inference is performed by sampling data from the
null distribution matching the estimated changepoint model
where each sample is drawn from a Gaussian and matches
in empirical means in each segment and in L2-norm (the sufficient statistics).
}
\details{
Currently the test is a 2-sided test.

The \code{test_func} should be a function that takes in \code{y} as the first
input and the output of \code{fit_method} as the second input, and \code{...}
as the third input.

If \code{sigma} is \code{NA}, then the function conditions on \code{.l2norm(y)}.
Otherwise, \code{sigma} represents the standard deviation of the marginal distribution
of \code{y}, asserting that each coordinate is independent and Gaussian.
}
