% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectiveModel.R
\name{selected_model_inference}
\alias{selected_model_inference}
\title{Selected model inference}
\usage{
selected_model_inference(y, fit_method, test_func, num_samp, sample_method,
  param = list(), cores = NA, verbose = T)
}
\arguments{
\item{y}{data of length n}

\item{fit_method}{function to fit changepoint model}

\item{test_func}{function to apply to each \code{y} and fitted model}

\item{num_samp}{number of desired samples from null distribution}

\item{sample_method}{either \code{hit_run} or \code{rejection}}

\item{param}{additional parameters for \code{sample_method} passed in as a list}

\item{cores}{number of cores}

\item{verbose}{boolean}
}
\value{
quantile
}
\description{
Fits a changepoint model using \code{fit_method} to \code{y}. The returned
model needs to have a function defined via the generic function
\code{binSeg::polyhedra}. Inference is performed by sampling data from the
null distribution matching the estimated changepoint model
where each sample is drawn from a Gaussian and matches
in empirical means in each segment and in L2-norm (the sufficient statistics).
}
